Fix multiple errors while using clang Change hooking not to depend on MSVC-specific casting rules.

+++ Source/FactoryGame/Public/FGProductionIndicatorInstanceManager.h
@@ -3,8 +3,9 @@
 
 #include "Components/HierarchicalInstancedStaticMeshComponent.h"
 #include "Buildables/FGBuildableFactory.h"
+//MODDING EDIT Add USceneComponent.h include
+#include "Components/SceneComponent.h"
 #include "FGProductionIndicatorInstanceManager.generated.h"
-
 
 /**
  * Instanced production indicator, used on factories to display their status.
@@ -44,11 +45,13 @@ private:
 	 * All instances managed, one instance list per indicator status.
 	 * Since these need to be attached to an actor to be rendered, our outer is used for this purpose.
 	 */
+	//MODDING EDIT: cast enum classes to underlying types because it is an syntax error on everything except MSVC
 	UPROPERTY()
-	UHierarchicalInstancedStaticMeshComponent* mInstanceComponents[ EProductionStatus::IS_MAX ];
+	UHierarchicalInstancedStaticMeshComponent* mInstanceComponents[ (uint8) EProductionStatus::IS_MAX ];
 
+	//MODDING EDIT: cast enum classes to underlying types because it is an syntax error on everything except MSVC
 	/** Handles from all the managed indicators. */
-	TArray< InstanceHandle* > mHandles[ EProductionStatus::IS_MAX ];
+	TArray< InstanceHandle* > mHandles[ (uint8) EProductionStatus::IS_MAX ];
 };
 
 //</CSS>
